+++ lib/generators/cms/component/error_tracking/airbrake/airbrake_generator.rb
@@ -22,16 +22,6 @@ module Cms
             template('airbrake.rb', File.join('config/initializers', 'airbrake.rb'))
           end
 
-          def update_local_custom_cloud_file
-            path = File.join(destination_root, 'config/custom_cloud.yml')
-
-            if File.exist?(path)
-              append_file(path) do
-                File.read(find_in_source_paths('custom_cloud.yml'))
-              end
-            end
-          end
-
           def add_deployment_notification
             unless options[:skip_deployment_notification]
               destination = 'deploy/after_restart.rb'
@@ -43,15 +33,6 @@ module Cms
               append_file(destination) do
                 File.read(find_in_source_paths('after_restart.rb'))
               end
-            end
-          end
-
-          def display_notice
-            if behavior == :invoke
-              log(:config,
-                'Please add your Airbrake API key to the section ' +
-                '"airbrake" in "config/custom_cloud.yml".'
-              )
             end
           end
         end