+++ lib/generators/cms/component/login_page/example/templates/migration.rb
@@ -8,12 +8,6 @@ class LoginPageExample < ::RailsConnector::Migration
       headline: 'Log in'
     )
 
-    reset_password_page = Obj.create(
-      _path: "#{configuration_path}/reset-password",
-      _obj_class: 'ResetPasswordPage',
-      headline: 'Reset Password'
-    )
-
     attributes = get_obj_class('Homepage')['attributes']
 
     attributes << {
@@ -22,18 +16,11 @@ class LoginPageExample < ::RailsConnector::Migration
       title: 'Login Page',
     }
 
-    attributes << {
-      name: 'reset_password_page',
-      type: :reference,
-      title: 'Reset Password Page',
-    }
-
     update_obj_class('Homepage', attributes: attributes)
 
     obj = Obj.find_by_path('/en')
     obj.update(
-      login_page: login_page['id'],
-      reset_password_page: reset_password_page['id'],
+      login_page: login_page['id']
     )
   end
 end