+++ lib/generators/cms/component/monitoring/newrelic/newrelic_generator.rb
@@ -40,25 +40,6 @@ module Cms
             end
           end
 
-          def update_local_custom_cloud_file
-            path = File.join(destination_root, 'config/custom_cloud.yml')
-
-            if File.exist?(path)
-              append_file(path) do
-                File.read(find_in_source_paths('custom_cloud.yml'))
-              end
-            end
-          end
-
-          def display_notice
-            if behavior == :invoke
-              log(:config,
-                'Please add your NewRelic API keys to the section ' +
-                '"newrelic" in "config/custom_cloud.yml".'
-              )
-            end
-          end
-
           private
 
           def append_template(source, destination)