+++ lib/infopark_kickstarter/rake/integration_task.rb
@@ -86,8 +86,6 @@ module InfoparkKickstarter
           'cms:widget:youtube:example',
           'cms:widget:vimeo',
           'cms:widget:vimeo:example',
-          'cms:widget:person',
-          'cms:widget:person:example',
           'cms:widget:diagram',
           'cms:widget:diagram:example',
           'cms:widget:slider',