+++ tmp/test_app/app/controllers/login_page_controller.rb
@@ -1,20 +1,25 @@
 class LoginPageController < CmsController
   def index
     if request.post?
-      contact = Infopark::Crm::Contact.authenticate(user_params[:login], user_params[:password])
-      self.current_user = User.new(contact.id)
+      if authenticate(user_params[:login], user_params[:password])
+        self.current_user = User.new
 
-      target = params[:return_to] || cms_path(@obj.homepage)
-      redirect_to(target, notice: 'You logged in successfully.')
+        target = params[:return_to] || cms_path(@obj.homepage)
+        redirect_to(target, notice: 'You logged in successfully.')
+      else
+        flash[:alert] = 'Log in failed. Please try it again.'
+      end
     elsif request.delete?
       self.current_user = nil
       redirect_to(cms_path(@obj.homepage), notice: 'You logged out successfully.')
     end
-  rescue Infopark::Crm::Errors::AuthenticationFailed, ActiveResource::ResourceInvalid
-    flash[:alert] = 'Log in failed. Please try it again.'
   end
 
   private
+
+  def authenticate(login, password)
+    login == 'root' && password == 'root'
+  end
 
   def user_params
     params[:user]