+++ tmp/test_app/app/models/reset_password_page.rb
@@ -1,6 +1,0 @@
-class ResetPasswordPage < Page
-  # Overriden method +show_breadcrumbs?+ from +Page+.
-  def show_breadcrumbs?
-    false
-  end
-end