+++ tmp/test_app/config/initializers/honeybadger.rb
@@ -1,10 +1,3 @@
-def honeybadger_config
-  config = YAML.load_file(Rails.root + 'config/custom_cloud.yml')
-  config['honeybadger'] || {}
-rescue Errno::ENOENT
-  {}
-end
-
 Honeybadger.configure do |config|
-  config.api_key = ENV['HONEYBADGER_API_KEY'] || honeybadger_config['api_key']
+  config.api_key = ENV['HONEYBADGER_API_KEY'] || ''
 end